models/requirements.txt
djangorestframework
requests
psycopg2